Messenger zxw

+++ Assets/ConnectApp/Utils/Sqlite/SQLiteDBManager.cs
@@ -37,7 +37,7 @@ namespace ConnectApp.Utils {
             this.m_Connection.InsertAll(data, extra: "OR REPLACE");
         }
 
-        public IEnumerable<DBMessageLite> QueryMessages(string channelId, int maxNonce = -1, int maxCount = 5) {
+        public IEnumerable<DBMessageLite> QueryMessages(string channelId, long maxNonce = -1, int maxCount = 5) {
             if (maxNonce == -1) {
                 return this.m_Connection.Table<DBMessageLite>().Where(message => message.channelId == channelId)
                     .OrderByDescending(message => message.nonce).Take(maxCount);
